small tweak

+++ tutorials/rm_pt2_working_dirs.py
@@ -4,6 +4,8 @@ dirs = [d for d in os.listdir(".") if os.path.isdir(d)]
 for d in dirs:
     dirs = [os.path.join(d,dd) for dd in os.listdir(d) if os.path.isdir(os.path.join(d,dd)) and ("master" in dd or "template" in dd or "worker" in dd)]
     for dd in dirs:
+        if "master_pp" in dd and "intro" in dd:
+            continue
         try:
             shutil.rmtree(dd,ignore_errors=True)
         except Exception as e:
